removes array allocation from HashUtil.ComputeHash instead using array parameters. updates ForestNodeBase to use casted GetHashCode to avoid boxing.

+++ libraries/Pliant/Forest/ForestNodeBase.cs
@@ -36,7 +36,7 @@
         private int ComputeHashCode()
         {
             return HashUtil.ComputeHash(
-                NodeType.GetHashCode(), 
+                ((int)NodeType).GetHashCode(), 
                 Location.GetHashCode(), 
                 Origin.GetHashCode());
         }
